:boom: Optimize directory structure
