--- src/api/association/association.interface.ts
@@ -1,8 +1,0 @@
-import { Types } from 'mongoose'
-
-export interface Association {
-  _id: Types.ObjectId
-  name: string
-  location: string
-  certificate: string
-}
