+++ src/dto/association-items.ts
@@ -1,7 +1,6 @@
 import { Expose, Type } from 'class-transformer'
-import { PaginationInfoDto } from '../pagination-info.dto'
-import { Association } from './association.interface'
-import { AssociationDto } from './association.dto'
+import { PaginationInfoDto } from './pagination-info'
+import { AssociationDto } from './association'
 
 /**
  * @openapi
@@ -25,5 +24,5 @@ export class AssociationItemsDto {
 
   @Expose()
   @Type(() => AssociationDto)
-  items!: Association[]
+  items!: AssociationDto[]
 }
